Will do sub string match of the instructor output with student output
The code server used to match student and instructor output.
If true then the code is correct.

The code server must do substring match. This is because student can write 'n' number of lines in the output stream. So the instructor code output must be in the student code output.