Functions added to compile and execute java code.
The first function compiles the code and returns error if any.
If there are no errors in the code then the next function is called
which executes and test the code with test cases. If the student code and instructor code
gives similar output then the function returns the code as correct code else as wrong code