Function modified to catch errors thrown by scilab
Actually when we execute scilab-cli command and test the file, it runs well
when there are no errors.
But if there are errors then the scilab subprocess does not exit.
It goes in the scilab-cli prompt, so can be only exited by entering exit
command.
So the user gets time out error instead of the actual error.
So that the user get the actual error, the user submitted code is
appended by the code which catches the error and writes into a file.
So if there is an error then the error file exists. Therefore error is read from
this file and displayed to the user.
Then the error file is deleted.